pyproject.toml
[project]
name = "workflow-py-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-ulid>=1.0.3",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "temporalio>=1.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
